feat: Improve image URL handling in ModuleTemplate
- Add `getImageURL` method to validate and retrieve the correct image URL
- Update the `Base` method to use the new `getImageURL` method
- Improve the error handling and logging in the `getImageURL` method

docs: Update the documentation for the `DownloadFile` and `CloneGitRepo` methods
- Add detailed descriptions of the parameters and functionality of these methods
- Improve the comments and structure of the documentation

ci: Bump the Dagger version in the `validateDaggerVersion` method
- Update the Dagger version to the latest stable version (v0.12.7)
