+++ .daggerx/templates/module/content.go.tmpl
