+++ .daggerx/templates/module/dagger.go.tmpl
@@ -2,11 +2,32 @@ package main
 
 import (
 	"context"
+	"fmt"
 	"strings"
 
 	"github.com/Excoriate/daggerverse/{{.module_name_pkg}}/internal/dagger"
 	"golang.org/x/mod/semver"
 )
+
+const (
+	// dockerVersionDefault specifies the default Docker version.
+	dockerVersionDefault = "24.0"
+)
+
+// getDockerInDockerImage returns the Docker-in-Docker image with the given version.
+//
+// If the version is not provided, it defaults to dockerVersionDefault.
+//
+// Example:
+//
+//	getDockerInDockerImage("20.10.17") => "docker:20.10.17-dind"
+func getDockerInDockerImage(version string) string {
+	if version == "" {
+		version = dockerVersionDefault
+	}
+
+	return fmt.Sprintf("docker:%s-dind", version)
+}
 
 // getDaggerInstallCMDByVersion returns the command to install the Dagger engine.
 //
@@ -15,8 +36,8 @@ import (
 //
 // Example:
 //
-//	getDaggerInstallCMDByVersion("v0.12.1")
-//	=> `cd / && DAGGER_VERSION="v0.12.1" curl -L https://dl.dagger.io/dagger/install.sh | DAGGER_VERSION="v0.12.1" sh`
+//	getDaggerInstallCMDByVersion("v0.12.7")
+//	=> `cd / && DAGGER_VERSION="v0.12.7" curl -L https://dl.dagger.io/dagger/install.sh | DAGGER_VERSION="v0.12.7" sh`
 func getDaggerInstallCMDByVersion(version string) string {
 	return strings.Join([]string{
 		"cd /",
@@ -31,7 +52,7 @@ func getDaggerInstallCMDByVersion(version string) string {
 // WithDaggerCLIAlpine sets up the Dagger CLI entry point for Alpine within the {{.module_name}}.
 //
 // Parameters:
-//   - version: The version of the Dagger Engine to use, e.g., "v0.12.1".
+//   - version: The version of the Dagger Engine to use, e.g., "v0.12.7".
 //
 // This method performs the following steps:
 //  1. Generates a shell command to install the Dagger CLI using the specified version.
@@ -58,7 +79,7 @@ func (m *{{.module_name}}) WithDaggerCLIAlpine(version string) *{{.module_name}}
 // WithDaggerCLIUbuntu sets up the Dagger CLI entry point for Ubuntu within the {{.module_name}}.
 //
 // Parameters:
-//   - version: The version of the Dagger Engine to use, e.g., "v0.12.1".
+//   - version: The version of the Dagger Engine to use, e.g., "v0.12.7".
 //
 // This method performs the following steps:
 //  1. Updates package lists and installs curl.
@@ -143,7 +164,7 @@ func (m *{{.module_name}}) validateDaggerVersion(dagVersion string) error {
 // container, and sets up the Docker service within the Dagger context.
 //
 // Parameters:
-//   - dagVersion: The version of the Dagger Engine to use, e.g., "v0.12.5".
+//   - dagVersion: The version of the Dagger Engine to use, e.g., "v0.12.7".
 //   - dockerVersion: The version of the Docker Engine to use, e.g., "24.0". This is optional.
 //
 // Returns:
